Removed checks on new certs in TLS cert reload action
It turned out that the validation performed on new certificates mostly
only complicated things and inhibited important use-cases without adding
significant value:

- It might be wanted that new certs have a shorter life span than old
certs due to policy changes
- New certs might have an issuer which is different from the old certs
- It is not necessary for certificates to have SAN for IP and hostname
and OID to be valid

Furthermore, when calling the reload endpoint, the old certificate file
has been already overwritten. If the update is then aborted for any
reason and the certificate file is left the way it is, the new
certificate will get active as soon as the node is restarted, which is
probably not wanted.

Also, the checks did not check the CA certificates even though these
also get reloaded. Thus, it was also possible to break the configuration
without the validation noticing it before.

Thus, for actually making the certificate update process safe of human
error, we'd need to have a more sophisticated mechanism which takes
place BEFORE the certificate files are overwritten. For now, we don't do
any validation and rely on the user to properly test and playbook the
cert update process.