Replaced GuiceHolder by non-singleton GuiceDependencies
Singletons are a very bad idea when we need to start-up several ES nodes
within one JVM, as we do in unit tests. This once again made problems
where the RepositoriesService suddenly believed it would be executing on
a non-master node even though it was executing on a master node.