Delete tenant_template when we detect that Kibana starts a migration
This is a hack for fixing
https://git.floragunn.com/search-guard/search-guard-kibana-plugin/-/issues/381

Instead of using a template, the Kibana plugin must now put the mapping
individually for each index. This is more compatible with Kibana index
migration. As there is no other way to remove the template in time, we
need to intercept the starting migration in the backend.