New test installer
This is a test installation script which is capable of starting a test cluster with just one command. The command will download all other necessary components, incl. OpenSearch or Elasticsearch.

The installation script is thoroughly commented and is intended as educational source on the configuration of Search Guard.

See merge request search-guard/search-guard-suite-enterprise!115