Metrics
Introduces a light-weight metrics mechanism for the Search Guard components.

The metrics can be retrieved using the `/_searchguard/component/_all/_health` API.

Whether a component collects metrics or not depends on the configuration of the component itself. Thus, metrics can be enabled for authentication, but disabled for authorization.

There are three levels:

* `NONE`: No metrics are collected
* `BASIC`: Default. A basic set of metrics is collected, which do not create any relevant overhead to the processing. Suitable for production systems.
* `DETAILED`: A more detailed set of metrics is collected. This might have a larger footprint; in many cases, that would also work for production. However, this should be tested.

See merge request search-guard/search-guard-suite-enterprise!200