Infrastructure for migration to new .searchguard index name
All internal Search Guard indices should be hidden indices. So far, the
searchguard index is not. Thus, it makes trouble when working with
wildcards.

This adds functionality to switch to a new index called .searchguard
during runtime. Basically, the ConfigurationRepository now looks for an
index called searchguard and for another one called .searchguard. As
soon as .searchguard exists, this is preferred over the old searchguard
index.

This does not add the actual migration code. This will be part of sgctl.