Removed transport client authentication
The transport client is deprecated and will be removed soon (or actually has been already removed); we should take the chance with FLX to remove transport client authentication for all versions, even for ES/OS versions which still support the transport client.

For those platforms, any try to authenticate with a transport client will fail.

This still keeps admin cert authentication for transport clients in order to be able to use sgadmin in legacy contexts.

See merge request search-guard/search-guard-suite-enterprise!222