Set explicit dependency on jakarta.activation and jakarta.activation-api
We are using many dependencies which use either directly or indirectly
jakarta.activation or jakarta.activation-api. Additionally, there is the
special situation that jakarta.activation contains all classes from
jakarta.activation-api without having it as a dependency (which would be
the usual way).

Thus, we often get different versions of jakarta.activation and
jakarta.activation-api via transitive dependencies which then leads to
the JAR hell issue in Elasticsearch.

So far, we mitigated this with many dependency exclusion rules, which
however got a bit out of hand - we has really quite a lot of
dependencies with exclusions on activation.

This commit solves this now by specifying jakarta.activation as a global
compile dependency. Additionally, activation-api is specified as a
global provided dependency (as it is provided by  jakarta.activation).
This should make other exclusions unnecessary.