Execute updateTenants() async to avoid blocking the management thread
The ConfigurationChangeListener onChange() method is executed by a
thread from the management thread pool. On one processor systems, this
thread pool will have a fixed size of 1 (one). Thus, any blocking call
on this thread has the danger of leading the node into a deadlock
because the management thread is unavailable to other operations which
may be spawned indirectly from the blocking operations.

Fixes #255