Corrected privilege support for _analyze action
Closes #240

See merge request search-guard/search-guard-suite-enterprise!652

The _analyze API is kind of a special thing: It allows to specify either
none or exactly one index. Most other actions behave differently: They
either expect exactly one index (e.g. _get) or 0 to n indices (e.g.
_search). In the latter case, the 0 index case means "search in all
indices". This is also already implemented this way in the privilege
evaluation logic.

The privilege evaluation logic did not consider the _analyze outlier,
though. Additionally, the _analyze action would return a single element
array with a null value for the `indices()` call if no index was
specified. This is a kind of malformed value and led to NPEs in the
privilege evaluation code.

We solve this by specially treating AnalyzeAction.Request objects (and
also its super class SingleShardRequest). For SingleShardRequest, we no
longer look at the `indices()` call. Rather we look at the `index()`
call and check for null values there. We still treat null values as a
reference to all indices first.

However, as AnalyzeAction.Request with a null index does not refer to
any index, we add special code in the
`ActionRequestIntrospector.reduceIndices()` code to just leave the
request untouched if there are privileges for the action.

Additionally, we add the `_analyze` API privilege to the `SGS_READ` action
group, as this can be considered as a pure read access action.

If a user wants to use the `_analyze` API with an index, the behavior is
straight forward: They need to have privileges for that particular
index.

If a user wants to use the `_analyze` API without an index, they just need
analyze privileges for *any* index. It does not matter for which index,
as the action actually does not operate on an index in that case.