Ignore string values for expires_in attribute in OIDC response
Some IDPs send the expires_in attribute as string value - in violation
to the standard. Before Search Guard would fail with a
NullPointerException in these cases. Now, it just ignores the string
values.

Generally, Search Guard does not use the expires_in attribute at the
moment - that is okay as it only uses the access token once. So, no
behavioral change will occur.